# apps/marketplace/internal/store/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package sqlc

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Credential struct {
	ID                    string
	ShopID                string
	ClientID              string
	SecretCiphertext      string
	Status                string
	CreatedAt             pgtype.Timestamptz
	RevokedAt             pgtype.Timestamptz
	AccessTokenCiphertext pgtype.Text
}

type DomainEvent struct {
	ID          string
	ShopID      string
	EventType   string
	AggregateID string
	Payload     []byte
	OccurredAt  pgtype.Timestamptz
}

type IdempotencyKey struct {
	ID             string
	CredentialID   string
	Operation      string
	Key            string
	ResponseStatus pgtype.Int4
	ResponseBody   []byte
	CreatedAt      pgtype.Timestamptz
	RequestHash    string
	State          string
	LockedUntil    pgtype.Timestamptz
}

type InventoryReservation struct {
	ID                string
	OrderID           string
	OrderItemID       string
	ProductID         string
	Quantity          int32
	Status            string
	ReservedAt        pgtype.Timestamptz
	CommittedAt       pgtype.Timestamptz
	ReleasedAt        pgtype.Timestamptz
	WarehouseID       pgtype.Text
	FulfilledQuantity int32
}

type Order struct {
	ID                     string
	OrderNumber            string
	ShopID                 string
	CustomerData           []byte
	ShippingAddress        []byte
	TotalAmount            int64
	Status                 string
	CreatedAt              pgtype.Timestamptz
	UpdatedAt              pgtype.Timestamptz
	PaymentReference       pgtype.Text
	PaidAt                 pgtype.Timestamptz
	PaymentStatus          string
	PaymentExpiresAt       pgtype.Timestamptz
	PaymentFailedAt        pgtype.Timestamptz
	PaymentFailureReason   pgtype.Text
	SellerDeadlineAt       pgtype.Timestamptz
	CancellationActor      pgtype.Text
	CancellationReason     pgtype.Text
	FulfillmentWarehouseID pgtype.Text
}

type OrderItem struct {
	ID          string
	OrderID     string
	ProductID   pgtype.Text
	Sku         string
	ProductName string
	Price       int64
	Quantity    int32
	Subtotal    int64
}

type Outbox struct {
	ID          string
	EventID     string
	PublishedAt pgtype.Timestamptz
	CreatedAt   pgtype.Timestamptz
}

type Package struct {
	ID          string
	OrderID     string
	Status      string
	CreatedAt   pgtype.Timestamptz
	UpdatedAt   pgtype.Timestamptz
	WarehouseID pgtype.Text
}

type PackageItem struct {
	PackageID   string
	OrderItemID string
	Quantity    int32
}

type Product struct {
	ID          string
	ShopID      string
	Sku         string
	Name        string
	Description string
	Price       int64
	Stock       int32
	Status      string
	CreatedAt   pgtype.Timestamptz
	UpdatedAt   pgtype.Timestamptz
	Category    string
}

type Shipment struct {
	ID                    string
	OrderID               string
	TrackingNumber        string
	ShippingProvider      string
	Status                string
	ShippedAt             pgtype.Timestamptz
	DeliveredAt           pgtype.Timestamptz
	PickupType            string
	CreatedAt             pgtype.Timestamptz
	PackageID             pgtype.Text
	DeliveryFailureReason pgtype.Text
	FailedAt              pgtype.Timestamptz
	ReturningAt           pgtype.Timestamptz
	ReturnedAt            pgtype.Timestamptz
}

type Shop struct {
	ID              string
	OwnerUserID     string
	Name            string
	Status          string
	CreatedAt       pgtype.Timestamptz
	UpdatedAt       pgtype.Timestamptz
	ProviderProfile string
}

type ShopScenario struct {
	ShopID                  string
	ApiSlowMs               int32
	ApiRandom500Probability int32
	ApiTimeoutProbability   int32
	ForceRateLimit          bool
	WebhookDuplicate        bool
	WebhookDelaySeconds     int32
	WebhookOutOfOrder       bool
	WebhookForceFailure     bool
	UpdatedAt               pgtype.Timestamptz
	ApiSlowProbability      int32
}

type SystemSetting struct {
	Key       string
	Value     []byte
	UpdatedAt pgtype.Timestamptz
}

type User struct {
	ID             string
	Email          string
	PasswordHash   string
	Role           string
	CreatedAt      pgtype.Timestamptz
	SessionVersion int32
}

type Warehouse struct {
	ID        string
	ShopID    string
	Code      string
	Name      string
	Status    string
	Address   []byte
	Priority  int32
	CreatedAt pgtype.Timestamptz
	UpdatedAt pgtype.Timestamptz
}

type WarehouseInventory struct {
	WarehouseID      string
	ProductID        string
	OnHandQuantity   int32
	ReservedQuantity int32
	CreatedAt        pgtype.Timestamptz
	UpdatedAt        pgtype.Timestamptz
}

type Webhook struct {
	ID               string
	ShopID           string
	Url              string
	SecretCiphertext string
	Enabled          bool
	SubscribedEvents []byte
	CreatedAt        pgtype.Timestamptz
}

type WebhookDelivery struct {
	ID            string
	WebhookID     string
	EventID       string
	Status        string
	AttemptCount  int32
	NextAttemptAt pgtype.Timestamptz
	CreatedAt     pgtype.Timestamptz
	DeliveredAt   pgtype.Timestamptz
	LeasedUntil   pgtype.Timestamptz
}

type WebhookDeliveryAttempt struct {
	ID              string
	DeliveryID      string
	Attempt         int32
	RequestHeaders  []byte
	ResponseStatus  pgtype.Int4
	ResponseHeaders []byte
	ResponseBody    pgtype.Text
	DurationMs      int32
	Status          string
	CreatedAt       pgtype.Timestamptz
}