# apps/marketplace/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Marketplace Simulator API
  version: 1.0.0
  description: |
    A provider-aware marketplace API. Order and fulfillment contracts are only
    available through `SHOPEE_LIKE` and `TOKOPEDIA_LIKE`; Generic is not an
    order-provider profile.

    Shared catalogue, warehouse, and webhook requests under `/api/v1` are HMAC-SHA256 signed using `METHOD + PATH + UNIX_TIMESTAMP + RAW_BODY`.
    Every mutating public operation across the shared, Shopee-like, and
    Tokopedia-like contracts require an `Idempotency-Key`. Concurrent use is
    claimed atomically, completed responses are replayed, and reusing a key for
    a different logical request returns `409`.

    Successful and rate-limited public responses expose `X-RateLimit-Limit`,
    `X-RateLimit-Remaining`, and `X-RateLimit-Reset` (Unix seconds). Public requests can return
    `503 MARKETPLACE_MAINTENANCE` when the administrator enables global maintenance mode.

    Webhooks are asynchronous. A 2xx response succeeds; all other responses and network failures
    retry at 30 seconds, 2 minutes, 10 minutes, and 30 minutes. Delivery attempts are immutable.
    Per-shop scenarios can inject slow responses, random 500s, timeouts, rate limits, duplicate,
    delayed, out-of-order, and failed webhook deliveries without affecting other shops.

    `SHOPEE_LIKE` shops additionally expose `/api/shopee/v1`. This is a distinct
...
    `X-Shopee-Signature`, where the signature is HMAC-SHA256 of
...

...
servers:
...
paths:
...
  /api/shopee/v1/orders:
    get:
      operationId: shopeeListOrders
      tags: [Orders]
      summary: List SHOPEE_LIKE orders with page-number pagination
      parameters: [{ $ref: '#/components/parameters/ShopeePartnerID' }, { $ref: '#/components/parameters/ShopeeTimestamp' }, { $ref: '#/components/parameters/ShopeeSignature' }, { name: page_no, in: query, schema: { type: integer, minimum: 1, default: 1 } }, { name: page_size, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }, { name: order_status, in: query, schema: { type: string } }, { name: time_from, in: query, description: Unix seconds, schema: { type: integer } }, { name: time_to, in: query, description: Unix seconds, schema: { type: integer } }]
      responses: { '200': { $ref: '#/components/responses/ShopeeResponse' }, '400': { $ref: '#/components/responses/ShopeeError' }, '401': { $ref: '#/components/responses/ShopeeError' }, '403': { $ref: '#/components/responses/ShopeeError' }, '429': { $ref: '#/components/responses/ShopeeError' } }
  /api/shopee/v1/products:
    get:
      operationId: shopeeListProducts
      tags: [Products]
      summary: List SHOPEE_LIKE catalogue items with Shopee-like terminology
      parameters: [{ $ref: '#/components/parameters/ShopeePartnerID' }, { $ref: '#/components/parameters/ShopeeTimestamp' }, { $ref: '#/components/parameters/ShopeeSignature' }, { name: page_no, in: query, schema: { type: integer, minimum: 1, default: 1 } }, { name: page_size, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }, { name: item_status, in: query, schema: { type: string, enum: [ACTIVE, INACTIVE] } }, { name: item_name, in: query, schema: { type: string } }]
      responses: { '200': { $ref: '#/components/responses/ShopeeResponse' }, '400': { $ref: '#/components/responses/ShopeeError' }, '401': { $ref: '#/components/responses/ShopeeError' }, '429': { $ref: '#/components/responses/ShopeeError' } }
  /api/shopee/v1/products/{id}:
    get:
      operationId: shopeeGetProduct
      tags: [Products]
      summary: Get one SHOPEE_LIKE catalogue item
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }, { $ref: '#/components/parameters/ShopeePartnerID' }, { $ref: '#/components/parameters/ShopeeTimestamp' }, { $ref: '#/components/parameters/ShopeeSignature' }]
      responses: { '200': { $ref: '#/components/responses/ShopeeResponse' }, '401': { $ref: '#/components/responses/ShopeeError' }, '404': { $ref: '#/components/responses/ShopeeError' } }
...
  /api/tokopedia/v202309/products/{id}:
    get:
      operationId: tokopediaGetProduct
      tags: [Tokopedia Like, Products]
      summary: Get one TOKOPEDIA_LIKE product
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }, { $ref: '#/components/parameters/TokopediaAppKey' }, { $ref: '#/components/parameters/TokopediaTimestamp' }, { $ref: '#/components/parameters/TokopediaSign' }, { $ref: '#/components/parameters/TokopediaAccessToken' }]
      responses: { '200': { $ref: '#/components/responses/TokopediaResponse' }, '400': { $ref: '#/components/responses/TokopediaError' }, '429': { $ref: '#/components/responses/TokopediaError' } }
...
  /api/v1/warehouses:
    get:
      operationId: listWarehouses
      tags: [Warehouses]
      summary: List fulfillment warehouses for the credential shop
      description: Warehouses are read-only from the public API. Orders reserve one active warehouse at creation; warehouse administration stays in the control plane.
      parameters: [{ $ref: '#/components/parameters/ClientID' }, { $ref: '#/components/parameters/Timestamp' }, { $ref: '#/components/parameters/Signature' }]
      responses: { '200': { $ref: '#/components/responses/RateLimitedWarehouseList' }, '401': { $ref: '#/components/responses/Error' }, '429': { $ref: '#/components/responses/RateLimited' }, '503': { $ref: '#/components/responses/Maintenance' } }
  /api/v1/warehouses/{id}:
    get:
      operationId: getWarehouse
      tags: [Warehouses]
      summary: Get one warehouse and its product inventory
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }, { $ref: '#/components/parameters/ClientID' }, { $ref: '#/components/parameters/Timestamp' }, { $ref: '#/components/parameters/Signature' }]
      responses: { '200': { $ref: '#/components/responses/RateLimitedWarehouse' }, '401': { $ref: '#/components/responses/Error' }, '404': { $ref: '#/components/responses/Error' }, '429': { $ref: '#/components/responses/RateLimited' }, '503': { $ref: '#/components/responses/Maintenance' } }
...
components:
  securitySchemes:
    ControlPlaneBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-session-token
  parameters:
    ClientID: { name: X-Client-Id, in: header, required: true, schema: { type: string } }
    Timestamp: { name: X-Timestamp, in: header, required: true, schema: { type: string, description: Unix seconds } }
    Signature: { name: X-Signature, in: header, required: true, schema: { type: string, description: Hex HMAC-SHA256 of METHOD + PATH + TIMESTAMP + raw body } }
    ShopeePartnerID: { name: X-Shopee-Partner-Id, in: header, required: true, schema: { type: string }, description: Credential client ID used as the provider partner ID. }
    ShopeeTimestamp: { name: X-Shopee-Timestamp, in: header, required: true, schema: { type: string, description: Unix seconds. } }
    ShopeeSignature: { name: X-Shopee-Signature, in: header, required: true, schema: { type: string, description: Hex HMAC-SHA256 of PARTNER_ID + PATH + TIMESTAMP + raw body. } }
    TokopediaAppKey: { name: app_key, in: query, required: true, schema: { type: string }, description: Partner Center app key (the control-plane credential client ID). }
    TokopediaTimestamp: { name: timestamp, in: query, required: true, schema: { type: integer, format: int64 }, description: Unix seconds; accepted within five minutes. }
    TokopediaSign: { name: sign, in: query, required: true, schema: { type: string }, description: Hex HMAC-SHA256 of APP_SECRET + PATH + sorted query excluding sign/access_token + raw body + APP_SECRET. }
    TokopediaAccessToken: { name: x-tts-access-token, in: header, required: true, schema: { type: string }, description: Access token returned only when the control-plane credential is created. }
...
  headers:
    RateLimitLimit: { description: Configured requests allowed in the current minute for this credential., schema: { type: integer } }
    RateLimitRemaining: { description: Remaining requests in the current minute. Zero on a 429 response., schema: { type: integer, minimum: 0 } }
    RateLimitReset: { description: Unix timestamp when the quota window resets., schema: { type: integer, format: int64 } }
  schemas:
...
    WarehouseInventory:
      type: object
      required: [product_id, sku, product_name, on_hand_quantity, reserved_quantity, available_quantity]
      properties:
        product_id: { type: string }
        sku: { type: string }
        product_name: { type: string }
        on_hand_quantity: { type: integer, minimum: 0 }
        reserved_quantity: { type: integer, minimum: 0 }
        available_quantity: { type: integer, minimum: 0 }
        updated_at: { type: string, format: date-time }
    Warehouse:
      type: object
      required: [id, shop_id, code, name, status, priority, created_at, updated_at]
      properties:
        id: { type: string }
        shop_id: { type: string }
        code: { type: string }
        name: { type: string }
        status: { type: string, enum: [ACTIVE, INACTIVE] }
        address:
          type: object
          description: Dispatch address configured for this fulfillment origin. The compatibility default warehouse may have an empty address until an operator sets it in the control plane.
          properties:
            address_line: { type: string, example: Jl. Raya Bekasi 10 }
            city: { type: string, example: Jakarta }
            postal_code: { type: string, example: "13910" }
          additionalProperties: true
        priority: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        inventory: { type: array, items: { $ref: '#/components/schemas/WarehouseInventory' } }
    WarehouseList:
      type: object
      required: [data]
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/Warehouse' } }
...
    Error:
      type: object
      properties: { error: { type: object, properties: { code: { type: string }, message: { type: string } } } }
  responses:
    ShopeeResponse: { description: Successful SHOPEE_LIKE response, headers: { X-Shopee-Api-Call-Limit: { schema: { type: string }, description: Calls used/current limit. }, X-Shopee-RateLimit-Reset: { schema: { type: integer }, description: Unix reset time. } }, content: { application/json: { schema: { type: object, required: [error, message, request_id, response], properties: { error: { type: string, example: '' }, message: { type: string, example: success }, request_id: { type: string }, response: { type: object } } } } } }
    ShopeeError: { description: SHOPEE_LIKE error, content: { application/json: { schema: { type: object, required: [error, message, request_id], properties: { error: { type: string, example: error_invalid_state }, message: { type: string }, request_id: { type: string } } } } } }
    TokopediaResponse: { description: Successful TOKOPEDIA_LIKE response, headers: { X-TTS-Api-Call-Limit: { schema: { type: integer } }, X-TTS-RateLimit-Reset: { schema: { type: integer, description: Unix reset time. } } }, content: { application/json: { schema: { type: object, required: [code, message, request_id, data], properties: { code: { type: integer, example: 0 }, message: { type: string, example: Success }, request_id: { type: string }, data: { type: object } } } } } }
    TokopediaError: { description: TOKOPEDIA_LIKE error, content: { application/json: { schema: { type: object, required: [code, message, request_id, data], properties: { code: { type: integer, example: 36000003 }, message: { type: string }, request_id: { type: string }, data: { type: object } } } } } }
    Error: { description: API error, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Maintenance: { description: Global maintenance is enabled., content: { application/json: { schema: { $ref: '#/components/schemas/Error' }, example: { error: { code: MARKETPLACE_MAINTENANCE, message: marketplace is under maintenance } } } } }
    RateLimited: { description: Credential quota is exhausted or the per-shop rate-limit scenario is enabled., headers: { X-RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }, X-RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }, X-RateLimit-Reset: { $ref: '#/components/headers/RateLimitReset' } }, content: { application/json: { schema: { $ref: '#/components/schemas/Error' }, example: { error: { code: RATE_LIMIT_EXCEEDED, message: rate limit exceeded } } } } }
...
    RateLimitedWarehouseList: { description: Warehouse list with credential quota headers., headers: { X-RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }, X-RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }, X-RateLimit-Reset: { $ref: '#/components/headers/RateLimitReset' } }, content: { application/json: { schema: { $ref: '#/components/schemas/WarehouseList' } } } }
    RateLimitedWarehouse: { description: Warehouse detail with credential quota headers., headers: { X-RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }, X-RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }, X-RateLimit-Reset: { $ref: '#/components/headers/RateLimitReset' } }, content: { application/json: { schema: { $ref: '#/components/schemas/Warehouse' } } } }
...